refactor: split create_recipe and run benchmark
